Get sender and issuer for alias and nft addresses
