+++ src/error.rs
@@ -258,10 +258,12 @@ pub enum Error {
     #[error("the semantic validation of a transaction failed with conflict reason: {} - {0:?}", *.0 as u8)]
     TransactionSemantic(ConflictReason),
     /// Local time doesn't match the time of the latest milestone timestamp
-    #[error("Local time {local_time} doesn't match the time of the latest milestone timestamp: {milestone_timestamp}")]
+    #[error(
+        "Local time {current_time} doesn't match the time of the latest milestone timestamp: {milestone_timestamp}"
+    )]
     TimeNotSynced {
         /// The local time.
-        local_time: u32,
+        current_time: u32,
         /// The timestamp of the latest milestone.
         milestone_timestamp: u32,
     },
